This fixes the auto-loaded lightboxes
These auto-loaded lightboxes are shown when a URL is loaded directly with the stem resource URL.
It moves the code for the lightbox divs into resource-lightbox component. The portal does not
use a stem-finder-result to render one of these lightboxes. So by moving the code into
resource-lightbox the code is shared by both.

There are still issues with this:
- scroll position: when a lightbox is opened from a card near the bottom of the page and then the page
  is reloaded. When the lightbox is closed it is confusing for the user.
- flash of content: when a lightbox is shown automatically by the portal the background content shows
  up quickly and then disappears.

[#150383056]