Refactor handling of short and long descriptions in portal-pages
This is designed to be backward compatible with older portals.
The new portal code will provide the actual (full) description, abstract, and description_for_teacher
When these are available portal_pages uses them to figure out the best thing to show.
The old portal code only provides the processed description, in this case portal_pages uses
this processed description everywhere.

[#150674559]
[#150414194]
[#150535877]
[#150675559]