this adds site-redesign to all of the built html
this needs more work, because locally a dev can add this the site-resign in
a configuration step. However, when copying the generated files up to to the
portal it needs to be manually added. This code tries to help with that, but it
doesn't handle all of the cases correctly.