Feature: Add Navigation testing page
Uses nested JSON to describe navigation structure.

PT:
[#157698255]
https://www.pivotaltracker.com/story/show/157698255

As a Concord Developer, I want all of my Portal rendering logic to be a separate application so I can more easily work on rendering/layout/views without the overhead of building the back end/databases. We will move left menu to portal-pages so that is part of the overall View App.