Add a simple 'examples'  directory in `dest` so that widgets can be tested from S3 distro.
[#158840067]
https://www.pivotaltracker.com/story/show/158840067

In order for QA to verify functionality, and in order for designers to preview upcoming portal pages changes, it should be easy for developers to include sample HTML files that showcase new features, and document the API changes.

Since the libraries are published to S3 anyway, this should be a small incremental improvement, or maybe just a case of improving developer documentation within the repo.