include css directly in portal-pages
This then uses the mini-css-extract-plugin to pull the css out during production build
All of the examples seem to be mostly working now.
This  works with webpack-serve, but the pages have errors because they are trying to
load portal-pages.css, and it doesn't exist in dev mode.
Some assets are still a problem with this config.
I think the css referenced assets are fine, but the javascript referenced assets are assumed to be
at /assets which doesn't exist yet.