Fix up start-portals server to work with new system
This removes the compiling of portal-pages.js and portal-pages.css from the start-portals server.
If you want to develop both the collection pages and the library at the same time you need to run
both start:prod and start-portals and use ResourceOverride to connect them. This is documented
in the readme changes.