This moves the recent collection logic up a level
Now the component will only be added to the page if the user is a teacher.
This means the request to the API will only be made if the user is a teacher.
Additionally the API request checks if the API url exists before trying to request it.
Without this change the the API was being requested with an undefined url which resulted
in an AJAX request to the current window location.