fix date filters on researcher report
The name property of the input was not set without this change.
The usage and details report buttons do full browser form submits.
So without the name property the browser will not send this field to
the server. The statistics at the top of the page were working properly
because they were based on a React state instead of the actual
input element in the DOM.

[#170675007]