Fix incorrect include when rendering materials-bins [#172782870]
The outer materials-bin component loaded the (very similarly named) collections display component instead of the wrapper component that loaded the collection.

The tests have been updated to load the collections and code was added to detect if a category did not have collections (categories can optionally only have sub-categories) as without the collections the render code broke.